requirements.txt
pandas==2.1.4
boto3
redshift-connector
sqlalchemy-redshift
python-dotenv
sqlalchemy
pydantic==2.5.0
pydantic-settings==2.1.0
psutil
pyarrow
